requirements.txt
telethon
httpx
Flask
psycopg2-binary
python-dotenv
gunicorn